LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/source-relevant-mutations-expected.txt



FAIL Neither the removing, moving, nor insertion steps for the source element track the previous 'next sibling' pointer when triggering relevant mutations for remaining child image elements of its old parent assert_equals: [img1] Img src attribute is chosen after third source is moved expected "data:,img1" but got "data:,s3"
